Veri tabanı anahtar değerleri long olarak değiştirildi. Süre değeri long olarak değiştirildi.

+++ app/src/main/java/com/example/kelimebulma/HomeActivity.java
@@ -88,7 +88,7 @@ public class HomeActivity extends AppCompatActivity {
                     if (timeMs >= mPuan.sureMs)
                         return;
 
-                puanDao.setSureMs(mPuan.id, (int) timeMs);
+                puanDao.setSureMs(mPuan.id, timeMs);
 
                 mTimeText.setText(GameHelper.getTimeString(timeMs));
                 break;
