+++ app/src/main/java/com/example/kelimebulma/dao/KullaniciDao.java
@@ -10,19 +10,19 @@ import com.example.kelimebulma.model.Kullanici;
 @Dao
 public interface KullaniciDao {
     @Query("SELECT kullanici_adi FROM kullanici WHERE id = :id")
-    String getKullaniciAdi(int id);
+    String getKullaniciAdi(long id);
 
     @Query("UPDATE kullanici SET kullanici_adi = :kullaniciAdi WHERE id = :id")
-    void setKullaniciAdi(int id, String kullaniciAdi);
+    void setKullaniciAdi(long id, String kullaniciAdi);
 
     @Query("SELECT sifre FROM kullanici WHERE id = :id")
-    String getSifre(int id);
+    String getSifre(long id);
 
     @Query("UPDATE kullanici SET sifre = :sifre WHERE id = :id")
-    void setSifre(int id, String sifre);
+    void setSifre(long id, String sifre);
 
     @Query("SELECT * FROM kullanici WHERE id = :id")
-    Kullanici getKullanici(int id);
+    Kullanici getKullanici(long id);
 
     @Query("SELECT * FROM kullanici WHERE kullanici_adi = :kullaniciAdi")
     Kullanici getKullanici(String kullaniciAdi);
